added a Context trait
