+++ parser/src/input_parser/file.rs
@@ -15,7 +15,7 @@
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
 use super::*;
-use crate::{SpannedToken, Token};
+use crate::{assert_no_whitespace, common::Context, SpannedToken, Token};
 use leo_errors::{ParserError, Result};
 
 use smallvec::smallvec;
