+++ parser/src/parser/file.rs
@@ -16,6 +16,7 @@
 
 use super::*;
 use crate::KEYWORD_TOKENS;
+use crate::common::Context;
 
 use leo_errors::{ParserError, Result};
 use leo_span::sym;
