+++ parser/src/parser/type_.rs
@@ -16,6 +16,7 @@
 
 use super::*;
 use leo_errors::{ParserError, Result};
+use crate::common::Context;
 
 use smallvec::smallvec;
